refactor: Changed overall response rate (ORR) from Double to Int and scaled it to interval [0,100]

+++ api/src/main/scala/de/dnpm/dip/mtb/query/api/MTBResultSet.scala
@@ -161,7 +161,7 @@ object MTBResultSet
     medications: Set[DisplayLabel[Coding[Medications]]],
     supportingAlteration: DisplayLabel[GeneAlteration],
     count: Int,
-    orr: Double,
+    orr: Int,  // Percent 0 - 100
     responseDistribution: Distribution[RECIST.Value],
     meanDuration: Double
   )
